# frontend/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Saknussemm",
    "description": "Post-OCR text correction of ALTO/PAGE XML files using LLM providers.",
    "version": "0.9.0"
  },
  "paths": {
...
    "/api/jobs": {
      "post": {
        "tags": ["jobs"],
        "summary": "Create Job",
        "description": "Upload ALTO/PAGE XML files and start a correction job.\n\nUpload concurrency is reserved by ``UploadAdmissionMiddleware``\nBEFORE the multipart body is read: by the time this handler runs,\nFastAPI has already parsed (and spooled) the files, so a slot check\nhere would fire after the full upload cost was paid. The whole\nhandler executes inside the middleware's reservation.",
        "operationId": "create_job_api_jobs_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_job_api_jobs_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/jobs/{job_id}/reviews": {
      "put": {
        "tags": ["review"],
        "summary": "Put Reviews",
        "description": "Record or replace judgements on lines of this job.\n\nIdempotent per line: sending the same line twice replaces its review\nrather than appending, so a reader who changes their mind is not fighting\nan append-only log. The timestamp is stamped here rather than trusted\nfrom the client \u2014 a review's date is a fact about the server.",
        "operationId": "put_reviews_api_jobs__job_id__reviews_put",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewBatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["review"],
        "summary": "Get Reviews",
        "operationId": "get_reviews_api_jobs__job_id__reviews_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/reviews/activity": {
      "post": {
        "tags": ["review"],
        "summary": "Review Activity",
        "description": "A reader interacted with the review UI; renew its idle timeout.",
        "operationId": "review_activity_api_jobs__job_id__reviews_activity_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/reviews/export": {
      "get": {
        "tags": ["review"],
        "summary": "Export Reviews",
        "description": "Export saved annotations and their source/candidate evidence.",
        "operationId": "export_reviews_api_jobs__job_id__reviews_export_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_create_job_api_jobs_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "geometric_pairing": {
            "type": "boolean",
            "title": "Geometric Pairing",
            "default": true
          }
        },
        "type": "object",
        "required": ["files", "provider", "api_key", "model"],
        "title": "Body_create_job_api_jobs_post"
      },
      "CreateJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "job_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Token"
          },
          "events_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Events Url"
          }
        },
        "type": "object",
        "required": ["job_id"],
        "title": "CreateJobResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LineReview": {
        "properties": {
          "page_id": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Page Id"
          },
          "line_id": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Line Id"
          },
          "verdict": {
            "$ref": "#/components/schemas/ReviewVerdict"
          },
          "transcription": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcription"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          }
        },
        "type": "object",
        "required": ["page_id", "line_id", "verdict"],
        "title": "LineReview",
        "description": "One reader's judgement on one line."
      },
...
      "ReviewBatch": {
        "properties": {
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/LineReview"
            },
            "type": "array",
            "maxItems": 2000,
            "title": "Reviews"
          }
        },
        "type": "object",
        "required": ["reviews"],
        "title": "ReviewBatch",
        "description": "Reviews arrive in batches: a reader works through a page, not a line."
      },
      "ReviewVerdict": {
        "type": "string",
        "enum": ["accepted", "refused", "transcribed"],
        "title": "ReviewVerdict",
        "description": "What the reader concluded about the engine's decision on this line."
      },
      "ReviewsResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/LineReview"
            },
            "type": "array",
            "title": "Reviews"
          }
        },
        "type": "object",
        "required": ["job_id", "reviews"],
        "title": "ReviewsResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}